Declare numerical variables as float by default

+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/visitor/LeilaVisitor.java
@@ -1105,7 +1105,7 @@ public class LeilaVisitor extends PrintVisitor {
 
     @Override
     public void visit(NumberType numberType) {
-        emitNoSpace("integer");
+        emitNoSpace("float");
     }
 
     @Override
